Fixed crash when source link was sent through ED2KLinks (introduced in 9435)
Two problems:
- ToLong behaves unexpectedly in wx 2.8 changing the argument even when returning false (fixed in wx 2.9)
- Cat check in AddLinksFromFile comes too late, invalid cat causes crash before in AddED2KLink
Also made cat type consistent (uint8).

See http://www.amule.org/amule/index.php?topic=17033.0

+++ trunk/src/DownloadQueue.h
@@ -265,11 +265,11 @@ public:
 	/**
 	 * Adds an ed2k or magnet link to download queue.
 	 */
-	bool	AddLink( const wxString& link, int category = 0 );
+	bool	AddLink( const wxString& link, uint8 category = 0 );
 
-	bool	AddED2KLink( const wxString& link, int category = 0 );
-	bool	AddED2KLink( const CED2KLink* link, int category = 0 );
-	bool	AddED2KLink( const CED2KFileLink* link, int category = 0 );
+	bool	AddED2KLink( const wxString& link, uint8 category = 0 );
+	bool	AddED2KLink( const CED2KLink* link, uint8 category = 0 );
+	bool	AddED2KLink( const CED2KFileLink* link, uint8 category = 0 );
 	bool	AddED2KLink( const CED2KServerLink* link );
 	bool	AddED2KLink( const CED2KServerListLink* link );
